+++ lib/assignments/data_source/user_data_source_impl.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
-import 'dart:io';
 
+import 'package:modu_3_dart_study/assignments/data_source/data_source_exception.dart';
 import 'package:modu_3_dart_study/assignments/data_source/user_data_source.dart';
 import 'package:modu_3_dart_study/assignments/data_source/user_model.dart';
 
@@ -11,7 +11,7 @@ class UserDataSourceImpl implements UserDataSource {
 
   @override
   Future<List<User>> getUsers() async {
-    final jsonString = await File(_usersPath).readAsString();
+    final jsonString = await getJsonFile(_usersPath);
     final jsonList = jsonDecode(jsonString) as List;
 
     return jsonList.map((jsonMap) => User.fromJson(jsonMap)).toList();
